In the ooxml branch, convert the formula stuff from using hssf.HSSFWorkbook to ss.Workbook, so that everything now works for XSSF too
git-svn-id: https://svn.apache.org/repos/asf/poi/branches/ooxml@648454 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/hssf/usermodel/HSSFWorkbook.java
@@ -548,7 +548,7 @@ public class HSSFWorkbook extends POIDocument implements org.apache.poi.ss.userm
      * Used by some of the more obscure formula and 
      *  named range things.
      */
-    public short getExternalSheetIndex(int internalSheetIndex) {
+    public int getExternalSheetIndex(int internalSheetIndex) {
     	return workbook.checkExternSheet(internalSheetIndex);
     }
